Code and testing for StandardMetric (note this is a factory were we will need to add the rest of the metrics that are currently commented out after ints are done)
